requirements.txt
cryptg
pillow
telethon